+++ packages/libs/src/sdk/examples/signed-request/package.json
