+++ packages/libs/src/sdk/examples/signed-request/tsconfig.json
